+++ tests/testthat/test-fn.R
@@ -161,13 +161,13 @@ test_that("fn_fmls<- and fn_fmls_names<- change formals", {
   expect_identical(fn_fmls(fn), pairlist(b = 1))
 })
 
-test_that("fn_fmls<- and fn_fmls_names<- handle primitive functions", {
-  fn_fmls(`+`) <- list(a = 1, b = 2)
-  expect_true(is_closure(`+`))
-  expect_identical(fn_fmls(`+`), pairlist(a = 1, b = 2))
-
-  fn_fmls_names(`+`) <- c("A", "B")
-  expect_identical(fn_fmls(`+`), pairlist(A = 1, B = 2))
+test_that("fn_ functions requires closures", {
+  msg <- "must be an R function, not a primitive function"
+  expect_error(fn_fmls(`+`), msg)
+  expect_error(fn_fmls_names(`+`), msg)
+  expect_error(fn_fmls_syms(`+`), msg)
+  expect_error(fn_fmls(`+`) <- list(a = 1, b = 2), msg)
+  expect_error(fn_fmls_names(`+`) <- c("A", "B"), msg)
 })
 
 test_that("assignment methods preserve attributes", {